backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pytest==7.4.3
pytest-django==4.7.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
celery==5.3.6
redis==5.0.1
eventlet==0.33.3 